requirements.txt
requests>=2.31.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
cryptography>=41.0.0
websockets>=12.0
